moving `CampaignFooter` to dedicated file
